refactor datacontract serializer
